+++ src/app/routes/routes.ts
@@ -1,0 +1,9 @@
+enum AppRoutes {
+  Root = '/',
+  Offer = '/offer/:id',
+  Login = '/login',
+  Favourites = '/favourites',
+  Everything = '/*',
+}
+
+export { AppRoutes };
