+++ src/components/favorites/index.ts
@@ -1,0 +1,2 @@
+export {FavoriteLocationItems} from './favorite-location-items';
+export {FavoritesList} from './favorites-list';
