+++ src/components/main-page/city-places-list.tsx
@@ -1,0 +1,51 @@
+import { FC } from 'react';
+import { IOffer, Offers } from '../../types/offers';
+import { PlaceCardMark } from '../place-card-mark';
+import { Link } from 'react-router-dom';
+import { capitalize } from '../../utils';
+
+const CityPlace: FC<IOffer> = ({offer}) => {
+
+  const {id, title, type, price, isPremium, rating} = offer;
+
+  return (
+    <article className="cities__card place-card">
+      {isPremium ? <PlaceCardMark /> : ''}
+      <div className="cities__image-wrapper place-card__image-wrapper">
+        <Link to={`/offer/${id}`}>
+          <img className="place-card__image" src='img/apartment-01.jpg' width="260" height="200" alt="Place image"/>
+        </Link>
+      </div>
+      <div className="place-card__info">
+        <div className="place-card__price-wrapper">
+          <div className="place-card__price">
+            <b className="place-card__price-value">&euro;{price}</b>
+            <span className="place-card__price-text">&#47;&nbsp;night</span>
+          </div>
+          <button className="place-card__bookmark-button button" type="button">
+            <svg className="place-card__bookmark-icon" width="18" height="19">
+              <use xlinkHref="#icon-bookmark"></use>
+            </svg>
+            <span className="visually-hidden">To bookmarks</span>
+          </button>
+        </div>
+        <div className="place-card__rating rating">
+          <div className="place-card__stars rating__stars">
+            <span style={{width: `${rating * 20}%`}}></span>
+            <span className="visually-hidden">Rating</span>
+          </div>
+        </div>
+        <h2 className="place-card__name">
+          <Link to={`/offer/${id}`}>{title}</Link>
+        </h2>
+        <p className="place-card__type">{capitalize(type)}</p>
+      </div>
+    </article>
+  );
+};
+
+export const CityPlacesList: FC<Offers> = ({offers}) => (
+  <div className="cities__places-list places__list tabs__content">
+    {offers.map((item) => <CityPlace offer={item} key={item.id}/>)}
+  </div>
+);
