+++ src/components/main-page/index.ts
@@ -1,0 +1,2 @@
+export {CitiesTabs} from './cities-tabs';
+export {CityPlaces} from './city-places';
