+++ src/components/offer/index.ts
@@ -1,0 +1,13 @@
+export {OfferFeatures} from './offer-features';
+export {OfferGallery} from './offer-gallery';
+export {OfferHost} from './offer-host';
+export {OfferInside} from './offer-inside';
+export {OfferMap} from './offer-map';
+export {OfferMark} from './offer-mark';
+export {OfferNearPlaces} from './offer-near-places';
+export {OfferPrice} from './offer-price';
+export {OfferRating} from './offer-rating';
+export {OfferReviews} from './offer-reviews';
+export {OfferTitle} from './offer-title';
+export {OfferReviewForm} from './offer-review-form';
+
