+++ src/components/offer/offer-map.tsx
@@ -1,0 +1,3 @@
+export const OfferMap = () => (
+  <section className="offer__map map"></section>
+);
