+++ src/components/offer/offer-mark.tsx
@@ -1,0 +1,5 @@
+export const OfferMark = () => (
+  <div className="offer__mark">
+    <span>Premium</span>
+  </div>
+);
