--- src/routes/routes.ts
@@ -1,6 +1,0 @@
-enum AppRoutes {
-  Root = '/',
-  Offer = '/offer/:id',
-}
-
-export {AppRoutes};
